requirements.txt
requests>=2.31.0
apache-airflow==2.6.3
apache-nifi==1.19.0
pyspark==3.5.0
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.4
pandas==2.0.3
plotly==5.18.0
scipy>=1.11.3
statsmodels>=0.14.0
scikit-learn>=1.3.2
kaleido==0.2.1
python-dotenv>=1.0.0
docker-compose>=1.29.2
apache-superset==2.1.0
sqlalchemy-bigquery>=1.6.1
matplotlib>=3.8.0
seaborn>=0.13.0
